+++ src/main/kotlin/net/pwall/json/schema/codegen/NestedClass.kt
@@ -1,0 +1,3 @@
+package net.pwall.json.schema.codegen
+
+class NestedClass(val constraints: Constraints, var className: String)