+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
+## [0.2] - 2020-08-23
+### Changed
+- `validation_string.mustache`: Added validations `uri`, `uri-reference`, `ipv4`, `ipv6`
+- `CodeGenerator`: improved output of array of object
+
 ## [0.1] - 2020-08-20
 ### Added
 - all files: initial version (copied from `json-kotlin-schema`)